split fix provider code into modules
